Feature(option): Add option to install for testing
Add help and force to test automatically.

+++ vscode/install.zsh
@@ -1,8 +1,21 @@
 #! /usr/local/bin/zsh
 
+setopt globdots
+local -A opthash
+zparseopts -D -A opthash -- -force -help -test
+
+if [[ -n "${opthash[(i)--help]}" ]]; then
+        echo "Add option --force to install without checking." && exit;
+fi
+
+if [[ -z "${opthash[(i)--force]}"  ]]; then
+        read Ans"?Your file will be overwritten(Y/n): "
+        if [[ $Ans != 'Y' ]]; then echo 'Canceled\n' && exit;
+fi;
+
 EXEPATH=$0:A:h
 VSPATH="~/Library/Application\ Support/Code/User"
 ln -sfnv ${EXEPATH}/settings.json ${VSPATH}/settings.json
 ln -sfnv ${EXEPATH}/keybindings.json ${VSPATH}/keybindings.json
 
-code --install-extension $(cat ${EXEPATH}/Pluginfile)
+if [[ -z "${opthash[(i)--test]}" ]]; then code --install-extension $(cat ${EXEPATH}/Pluginfile) ; fi;
